# app/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Yüz Tanıma ile Yoklama Sistemi API",
    "description": "Yüz tanıma teknolojisi kullanarak yoklama almayı ve duygu analizi yapmayı amaçlayan bir sistemin API'si",
    "version": "1.0.0"
  },
  "host": "localhost:5000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\""
    }
  },
...
  "paths": {
...
    "/api/teachers": {
      "get": {
        "tags": [
          "👨‍🏫 Öğretmenler"
        ],
        "summary": "Tüm öğretmenleri listele",
        "description": "Sistemdeki tüm öğretmenleri listeler",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Başarılı",
            "schema": {
              "type": "object",
              "properties": {
                "items": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                },
                "page": {
                  "type": "integer",
                  "example": 1
                },
                "per_page": {
                  "type": "integer",
                  "example": 10
                },
                "total": {
                  "type": "integer",
                  "example": 2
                },
                "pages": {
                  "type": "integer",
                  "example": 1
                }
              }
            }
          },
          "401": {
            "description": "Yetkisiz erişim",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Yetkisiz erişim."
                }
              }
            }
          },
          "500": {
            "description": "Öğretmenler listelenemedi",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Öğretmenler listelenemedi."
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "👨‍🏫 Öğretmenler"
        ],
        "summary": "Yeni öğretmen oluştur",
        "description": "Sisteme yeni bir öğretmen ekler",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Öğretmen bilgileri",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "email",
                "password",
                "first_name",
                "last_name",
                "department"
              ],
              "properties": {
                "email": {
                  "type": "string",
                  "example": "ogretmen@example.com"
                },
                "password": {
                  "type": "string",
                  "example": "sifre123"
                },
                "first_name": {
                  "type": "string",
                  "example": "Ahmet"
                },
                "last_name": {
                  "type": "string",
                  "example": "Yılmaz"
                },
                "department": {
                  "type": "string",
                  "example": "Bilgisayar Mühendisliği"
                },
                "title": {
                  "type": "string",
                  "example": "Dr. Öğr. Üyesi"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Öğretmen başarıyla oluşturuldu",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Öğretmen başarıyla oluşturuldu."
                },
                "teacher": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Email zaten kullanılıyor",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Bu e-posta adresi zaten kullanılıyor."
                }
              }
            }
          },
          "500": {
            "description": "Öğretmen oluşturulamadı",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Öğretmen oluşturulamadı."
                }
              }
            }
          }
        }
      }
    },
    "/api/teachers/{teacher_id}": {
      "get": {
        "tags": [
          "👨‍🏫 Öğretmenler"
        ],
        "summary": "Öğretmen detayını getir",
        "description": "Belirli bir öğretmenin detaylarını getirir",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "teacher_id",
            "in": "path",
            "description": "Öğretmen ID",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Başarılı",
            "schema": {
              "type": "object"
            }
          },
          "404": {
            "description": "Öğretmen bulunamadı",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Öğretmen bulunamadı."
                }
              }
            }
          },
          "500": {
            "description": "Sunucu hatası",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Sunucu hatası."
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "👨‍🏫 Öğretmenler"
        ],
        "summary": "Öğretmen bilgilerini güncelle",
        "description": "Belirli bir öğretmenin bilgilerini günceller",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "teacher_id",
            "in": "path",
            "description": "Öğretmen ID",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Güncellenecek öğretmen bilgileri",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "yeni_ogretmen@example.com"
                },
                "first_name": {
                  "type": "string",
                  "example": "Yeni Ahmet"
                },
                "last_name": {
                  "type": "string",
                  "example": "Yeni Yılmaz"
                },
                "department": {
                  "type": "string",
                  "example": "Elektrik Elektronik Mühendisliği"
                },
                "title": {
                  "type": "string",
                  "example": "Prof. Dr."
                },
                "password": {
                  "type": "string",
                  "example": "yeni_sifre123"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Öğretmen başarıyla güncellendi",
            "schema": {
              "type": "object"
            }
          },
          "404": {
            "description": "Öğretmen bulunamadı",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Öğretmen bulunamadı."
                }
              }
            }
          },
          "500": {
            "description": "Öğretmen güncellenemedi",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Öğretmen güncellenemedi."
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "👨‍🏫 Öğretmenler"
        ],
        "summary": "Öğretmeni sil",
        "description": "Belirli bir öğretmeni sistemden siler",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "teacher_id",
            "in": "path",
            "description": "Öğretmen ID",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Öğretmen başarıyla silindi",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Öğretmen başarıyla silindi."
                }
              }
            }
          },
          "400": {
            "description": "Öğretmenin aktif dersleri var",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Öğretmenin aktif dersleri var. Önce dersleri silin veya başka bir öğretmene atayın."
                }
              }
            }
          },
          "403": {
            "description": "Bu işlem için admin yetkisi gerekli",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Bu işlem için admin yetkisi gerekli."
                }
              }
            }
          },
          "404": {
            "description": "Öğretmen bulunamadı",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Öğretmen bulunamadı."
                }
              }
            }
          },
          "500": {
            "description": "Öğretmen silinemedi",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Öğretmen silinemedi."
                }
              }
            }
          }
        }
      }
    },
...
  }
} 